Convert to full typescript
