+++ .eslintrc.yml
@@ -1,2 +1,28 @@
-extends:
-  - plugin:@lifeomic/node/recommended
+{
+  "overrides": [
+    {
+      "files": ["**/*.js"],
+      "extends": [
+          "plugin:@lifeomic/eslint-plugin-node/recommended"
+      ]
+    },
+    {
+      "files": ["**/*.ts"],
+      "extends": [
+          "plugin:@lifeomic/eslint-plugin-node/recommended",
+          "eslint:recommended",
+          "plugin:@typescript-eslint/eslint-recommended",
+          "plugin:@typescript-eslint/recommended"
+      ],
+      "parser": "@typescript-eslint/parser",
+      "plugins": ["@typescript-eslint"],
+      "rules": {
+        "@typescript-eslint/no-explicit-any": "off",
+        "@typescript-eslint/ban-ts-ignore": "off",
+        "@typescript-eslint/explicit-function-return-type": "off",
+        "@typescript-eslint/no-non-null-assertion": "off",
+        "@typescript-eslint/no-var-requires": "off"
+      }
+    }
+  ]
+}
