+++ .gitignore
@@ -1,2 +1,5 @@
 node_modules/
 .nyc_output/
+
+src/**/*.js
+test/**/*.js
