+++ package.json
@@ -11,13 +11,14 @@
     "url": "https://github.com/lifeomic/graphql-resolvers-xray-tracing.git"
   },
   "scripts": {
-    "lint": "eslint .",
+    "lint": "eslint . -f codeframe",
     "test": "nyc ava",
-    "posttest": "yarn lint",
-    "coverage": "nyc report --reporter=text-lcov > ./.nyc_output/lcov.info"
+    "posttest": "yarn lint && tsc --noEmit",
+    "coverage": "nyc report --reporter=text-lcov > ./.nyc_output/lcov.info",
+    "prepublishOnly": "yarn tsc"
   },
   "main": "src/traceResolvers.js",
-  "types": "src/traceResolvers.d.ts",
+  "types": "src/traceResolvers.ts",
   "engines": {
     "node": ">=10"
   },
@@ -33,6 +34,10 @@
   "devDependencies": {
     "@lifeomic/eslint-plugin-node": "^2.0.1",
     "@types/graphql": "^14.0.7",
+    "@types/promise-retry": "^1.1.3",
+    "@types/uuid": "^8.3.0",
+    "@typescript-eslint/eslint-plugin": "^2.33.0",
+    "@typescript-eslint/parser": "^2.33.0",
     "ava": "^3.14.0",
     "aws-xray-sdk-core": "^3.2.0",
     "coveralls": "^3.1.0",
@@ -43,6 +48,9 @@
     "nyc": "^15.1.0",
     "promise-retry": "^2.0.1",
     "sinon": "^9.2.2",
+    "source-map-support": "^0.5.19",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3",
     "uuid": "^8.3.2"
   },
   "nyc": {
@@ -59,13 +67,20 @@
   "ava": {
     "verbose": true,
     "files": [
-      "test/**/*.test.js"
+      "test/**/*.test.ts"
+    ],
+    "require": [
+      "ts-node/register",
+      "source-map-support/register"
+    ],
+    "extensions": [
+      "ts"
     ]
   },
   "publishConfig": {
     "access": "public"
   },
   "peerDependencies": {
-    "aws-xray-sdk-core": "^2.0.0"
+    "aws-xray-sdk-core": "^2.0.0 || ^3.0.0"
   }
 }
